FIX backport fix for lack of HTMLPurifier breakage
If HTMLPurifier is missing, a fatal error is giving for a missing class.
This results in comments essentially not working. The fix is simple and
has already been done in 2.0 - as the package is optional we can simply
check if the classname exists before trying to use it, or do nothing.